made moist uint16

+++ lib/sensors/structs/sensorReadings.h
@@ -5,7 +5,7 @@ struct SensorReading
     int16_t airTemp;
     int16_t airHumidity;
     int16_t soilTemp;
-    uint8_t soilMoisture;
+    uint16_t soilMoisture;
     uint8_t light : 1;
     uint16_t lightLux;
 } __attribute__((packed));
